requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
dataclasses==0.8
Flask==2.0.3
idna==3.3
importlib-metadata==4.8.3
interchange==2021.0.4
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
monotonic==1.6
packaging==21.3
pansi==2020.7.3
py2neo==2021.2.3
pyahocorasick==1.4.4
Pygments==2.11.2
pyparsing==3.0.7
pytz==2021.3
requests==2.27.1
six==1.16.0
typing_extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
zipp==3.6.0
